Extract classic directive lowering
